requirements.txt
requests
pymongo
python-dotenv
google-genai
